requirements.txt
# Supported Python runtime: >=3.10,<3.13 (see Documentation/RUNTIME_GUARDRAILS.md)
# Last updated: 2026-04-19 (Schema v4 readiness: pandas_market_calendars pin for schedule-aware freshness)
# Observability note: provision Prometheus, Alertmanager, Grafana, Loki, and Alloy under tools/observability or via scripts/install_observability_stack.ps1 rather than pip.
# Unfixed: pygments 2.19.2 CVE-2026-4539 (local-only ReDoS, no fix released); ignored in CI via --ignore-vuln-id.
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
alpha_vantage==3.0.0
annotated-types==0.7.0
anthropic==0.68.0
anyio==4.11.0
arch==8.0.0
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.2; python_version < "3.11"
contourpy==1.3.3; python_version >= "3.11" and python_version < "3.13"
curl_cffi==0.15.0
cryptography==46.0.7
cycler==0.12.1
deepdiff==8.6.2
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.128.6
fonttools==4.61.1
frozendict==2.4.6
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
kiwisolver==1.4.9
lxml==6.0.2
lxml-html-clean==0.4.4
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
multidict==6.7.1
multitasking==0.0.12
networkx==3.4.2; python_version < "3.11"
networkx==3.6.1; python_version >= "3.11" and python_version < "3.13"
numpy==2.4.0; python_version >= "3.11" and python_version < "3.13"
numpy==1.26.4; python_version < "3.11"
ollama==0.6.0
# OpenBB stack intentionally excluded from secure default profile due
# python-multipart constraint (<0.0.21) in current openbb-core release.
# If OpenBB is re-enabled, revisit CVE posture and dependency pins first.
packaging==26.0
pandas==2.3.3
pandas-datareader==0.10.0
pandas_market_calendars>=5.3.2,<6
patsy==1.0.2
peewee==3.18.3
pillow==12.1.1
platformdirs==4.9.2
pluggy==1.6.0
pre-commit==4.5.1
protobuf==6.33.5
propcache==0.4.1
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
hypothesis==6.151.9
pytest==9.0.0
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
PyJWT[crypto]==2.12.0
requests==2.33.0
scikit-learn==1.7.2
scipy==1.15.3; python_version < "3.11"
scipy==1.16.3; python_version >= "3.11" and python_version < "3.13"
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
statsmodels==0.14.6
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
uvicorn==0.41.0
websockets==15.0.1
yarl==1.22.0
yfinance==0.2.66
